pyproject.toml
[project]
name = "testing-agent-explorer"
version = "0.2.0"
description = "Automated mobile app explorer that builds state graphs via PUCT + Go-Explore"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.5.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "hypothesis>=6.98.0",
    "Pillow>=10.2.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
explorer = "explorer.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]
